Remove openstack-101 from status and meetbot
In discussions at the last PTG[1] and in subsequent First Contact
SIG meetings[2] we've decided that the #openstack-101 channel
should be removed and rolled into the #openstack-dev channel.

Why? Good qustion:
 - No one is good at sitting in and watching the #openstack-101
   channel outside of OUI training periods.
 - It would be better for new contributors to ask questions in
   a channel that people with the answers already activly monitor.
 - 101 is an American-ism, that not everyone in the world would
   understand.

This means that during OUI trainings there will be more noise in
openstack-dev. However, trainings mostly happen the weekend
before a summit, so at a very quiet time. And would mean trainees
would already have practice and now where to come for questions.

There is another patch[3] which will remove openstack-101 from
the accessbot.

Finally, as per renaming instructions[4], once these patches
land, a redirect will need to be created, the topic updated and
current residents of the channel will need to be informed to
relocate.

[1] https://etherpad.openstack.org/p/FC_SIG_Rocky_PTG
[2] https://wiki.openstack.org/wiki/First_Contact_SIG#Meeting_Agenda
[3]
[4] https://docs.openstack.org/infra/system-config/irc.html#renaming-an-irc-channel

Change-Id: I24710f24e7e640aabbf22353c25c24d2a8af9287