Add statusbot and meetbot to #openstack-placement
'#openstack-placement' is a new IRC channel for disussing the
placement service. It has been extracted from openstack-nova to make
that channel less noisy. We would like statusbot and meetbot in the
channel so that we can be aware of status and have meetings local to the
channel if necessary.

The depends on are the project-config changes for accessbot and
gerritbot.

Depends-On: https://review.openstack.org/559768
Change-Id: I0002c2c753b04c8cc02bd5ac1a1ab528bf775591