Update to latest puppet-python
Update to the latest version of this module

Thre is a conflict with the current virtualenv installation.  We are
making a 'virtualenv' package with a provider of openstack_pip in the
current code, but this conflicts with the 'virtualenv' package as
provided by puppet-python.

Use puppet-python to ensure python-dev, pip and virtualenv packages
are installed.  Once they are, we can use pip to install the latest
virtualenv as specified.

Change-Id: Iec329f12c1c6620210562e872c649d1cc715de51