Update to nodejs 6.x on etherpad.o.o
In order to upgrade to current versions of the Etherpad service, we
need newer nodejs than is available on Ubuntu Trusty. Switch to 6.x
which is the current default for the puppet-etherpad_lite module and
is also what we've tested on etherpad-dev.openstack.org. Switching
between Ubuntu and Nodesource package sources does not result in an
immediate package upgrade, so this must be performed manually before
we change the eplite_version parameter to something newer.

Change-Id: Ied9e5bd7ffa16f1832d3e1e26d0886de67f98f72