Update etherpad to 1.6.5
Etherpad releases 1.6.4 and 1.6.5 include some security updates as
well as a number of bug fixes over the random post-1.5.0 commit
we've pinned to for a while. Since it appears upstream is now
tagging releases with some regularity we can try pinning to specific
tags again.

Note that the vulnerabilities fixed in 1.6.4+ are either too new or
otherwise non-impacting for the version, configuration and use case
under which we're deploying in production, but this update is good
hygiene anyway.

Change-Id: Idc0e0c3a6d298aad2e41772e249c1a167c88559a