Fix the gerrit-host used by jeepyb config
This template was previously set to use the local hosts fqdn for the
jeepyb gerrit-host config item. Now that we've switched to indexed hosts
for review.o.o that ends up as review01.openstack.org.

This would work fine for network connectivity to the ssh server and
otherwise we wouldn't notice except that this is the value we write into
the .gitreview files. To make them more accurate switch to
review.openstack.org as the explicit value.

Note I considered making this configurable but we already assume so many
other openstack specific values in this file I don't think its super
necessary.

Change-Id: I065af01b2793639453562d924e58ba7c769dcd38