Pin /opt/ansible as source of openstack inventory
We clone ansible to /opt/ansible and use it as the source of our
openstack inventory script. Newer ansible has renamed this script so we
need to migrate to the new thing. Until we are ready to do that pin to
an older version of ansible that has the script at the old location.

Change-Id: I2084601b8f2f3629205b3c2c415bc1ad793226b0