Switch to using new ansible openstack inventory script
There were updates upstream in ansible to rename this script due to
import issues. Additionally this switches us from using shade to
openstacksdk to get the inventory contents dynamically.

Note that we ensure the old file is absent prior to adding the new file
to avoid a race where we'll have two dynamic inventory scripts providing
the same functionality.

Change-Id: I76b1099bf0cf3bfead17f96e456cdce87d0e8a49