Add helga-grafana to grafana for IRC alerts
Helga [1] is an extensible chatbot which has good support for adding
webhook handlers and relaying messages to irc.

helga-grafana [2] is a plugin which can take the webhook messages
(i.e. a POST) coming from the alerts produced by the new version 5.X
grafana install and send it to IRC.

This configures a minimal helga service.  It listens locally on port
8999 and relays alerts from grafana to #openstack-infra as
openstackgrafana.

There is still more to do with puppeting the configuration in grafana
and of course writing sensible alerts from statsd/graphite data (via
grafyaml), but it's a start!

[1] https://github.com/shaunduncan/helga
[2] https://pypi.org/project/helga-grafana/

Change-Id: I6d78a6f9eb84e188b75861ed5c80af7bedca6adc