Increase apache mirror cache to 70GB
We expect at least 100GB of disk available for the apache mirror caches
we run in each region. However, due to htcacheclean lagging apache's
ability to cache we cleaned up to 50GB to give us breathing room. Since
then we have reorganized the cache layout so that htcacheclean can
operate more quickly and don't need as much breathing room.

Increase this cache size value up to 70GB which still gives us a
conservative ~30GB or so of breathing room. We can continue to increase
this number if htcacheclean keeps up.

Change-Id: I8a13a5876cea930c5d981af2c7cd6ab36ba867fb