Switch ansible-lint to using multiple -x options
ansible-lint 3.4.22 seems to have introduced a change in behaviour
in how it parses the excludes (-x) back in March[1]. However, it
seems that the new release never got pushed to PyPi until now.

This has resulted in the linting jobs failing, so switching the
format to be the new one instead of pinning to an older release.

[1]: https://github.com/willthames/ansible-lint/commit/96d0a0bc

Change-Id: I446e734364298e90a64063e811343aa11430765f