Add debian-security mirror
As a follow-on to Id918597aecc5114791811c7f30b6109da6a8d0af, re-add
the debian-security mirror.  We have problems with transparent caches
getting to http://security.debian.org, and it doesn't have https, so
we really need this cached locally for reliable operation.

This adds debian-security as a separate mirror due to the conflicts we
saw in the prior change.

The keys are updated for stretch (check against
https://ftp-master.debian.org/keys.html)

Change-Id: Ic33bca431ae19b44e5bc2938de425cc68785b331