Fix MariaDB proxy backend details
There are two things that need fixing with the MariaDB repo proxy
caching. First it is actually the downloads.mariadb.com site not the
.org that serves the repos. Second they do relative path redirects for
when you miss /'s this means we have to share the same prefix as the
backend on our frontend so that relative redirects work.

This means using /MariaDB proxied to
https://downloads.mariadb.com/MariaDB.

Change-Id: If35617a8c883c49121d8833de180b16f7a0bc2ef