More minor debian-security mirror fixes
As a follow-on to Ic33bca431ae19b44e5bc2938de425cc68785b331
... (reprepro configuration really is quite inscrutable outside "try
it and see if it works")

The update is just called "debian-security", no codename.

We want to use the debian-security key to validate the gpg signatures,
update that.

Use the "updates" suite, which seems to be where debian-security is

Add the symlink for mirror

Change-Id: Ice2a06f4274cdad766f427df70d9fe3e92f62e8c