Upgrade to LimeSurvey 3.15.6
LimeSurvey releases early and often (yay!) but as a result we're
rather many versions behind just in the time it took us to finalize
our automation. Upgrade to the latest tagged version now. Also
switch to downloading from limesurvey.org instead of github.com as
the latter seems to only house a (delayed) mirror of their releases.
The updated location also implies slightly different file naming, so
a few of the Puppet execs dealing with retrieval and unpacking that
file have to be adjusted to match.

Change-Id: Ib6faa4cedcef6685f6992d61de9a29cd72f05aa2