Stop installing puppet with puppet
Currently, the openstack_project::server class unconditionally ensures
the 'puppet' package is installed. As long as we're using puppet 3, this
is a noop: if puppet is running, then it's already installled. However,
if we have installed puppet 4 from the puppetlabs repositories, then the
puppet 4 package is actually called 'puppet-agent', and reinstalling the
'puppet' package from the distro repositories will clobber it. We could
work around this in two ways: one is to add logic to the
openstack_project::server class to manage the correct package name and
correct package version based on a parameter or the current state of the
system, and the other is to stop managing it entirely from puppet. Since
we're already managing it from ansible, let's go with the
less-puppet-code option.

Change-Id: I32c02302a5009dcd3a54fae7fb346f2214be886a