Fix template variables
We almost never define $serveraliases which means that it's always going
to be empty. Puppet 4 is more strict about types and converting undef to
nil so this change makes sure that an undefined variable doesn't
accidentally get evaluated as existing and cause an invalid apache
config.

In the places where we're already just comparing @serveraliases to nil,
just leave them alone. Nil should be the expected value for undefined
variables going forward in puppet 4, so as long as it's working in
puppet 3 let's keep it.

Change-Id: I68ea8e1fde5d422f821453ff618d827e3cbe1c43