Add proxy cache for PyPI
Investigate using apache proxy caching for PyPI instead of a full
mirror. Make the /pypi path proxy directly to pypi.org upstream instead
of to the rackspace AFS mirror.

We have to use the substitute module (already enabled) to rewrite the
embedded links in the html content returned, because instead of relative
links to something under /pypi which get redirected, the CDN links are
direct.

Change-Id: Id68ff2914ba7fb588092cff113e8d1536f7d5a56