Serve pypi proxy cache from port 80
Oh easy_install.

In the configure-mirrors zuul-jobs role we write out ~/.pydistutils.cfg.
One of the things we do is:

  allow_hosts = {{ mirror_fqdn }}

which results in things like

  allow_hosts = mirror.dfw.rax.openstack.org

True story though - that doesn't match port 8080, so it breaks
easy_install if we set

  index_url = http://mirror.dfw.rax.openstack.org:8080/pypi/simple

Is there anything easy_install does right? No?

We should probably completely rethink how we're plumbing mirror info
into our nodes, but it is non-trivial to get a port into that file and
various OpenStack projects are currently blocked on the pypi mirror
being out of date.

Serve the pypi proxy content from port 80.

Change-Id: I518cf55fc6efe7db9493e4f5ec3601cb67f54999