Add beaker tests for firehose
firehose uses a collection of multiple related modules and it's
cumbersome to add tests to all of them, but this patch can demonstrate
that all of them are working.

This shouldn't be merged, since tests to any of the individual modules
won't cause these tests to run. This is just useful to show in general
that firehose will work for puppet 4.

Change-Id: I0d6bb9f82ab319788536acbf219f9a35ed9c6890