WIP Use ansible for openstack_project::server
Here we go folks.

Now that we've got base server stuff rewritten in ansible, remove the
old puppet versions and start running the base playbook everywhere.

Update the update_puppet_version playbook to do both 3 and 4 and set an
ansible variable to control which hosts get which version. Also, add a
group (that currently is just whatever isn't bridge) that defines which
hosts we run the puppet script on.

Depends-On: https://review.openstack.org/588326
Change-Id: I5c82fe6fd25b9ddaa77747db377ffa7e8bf23c7b