WIP Move base activities into ansible
Here we go folks.

As part of updating our config management approach, we need to move the
functionality provided by openstack_project::server into Ansible.

Start by making a couple of roles and a playbook and run that playbook
before we run any of the puppet.

The base roles are split into two, one for setting up additional repos
and one for actually doing the work. This lets us use handlers to run an
apt-get update after updating sources.list files.

The corresponding puppet has been removed so that it's clear for the
purposes of review what has happened.

Change-Id: I5c82fe6fd25b9ddaa77747db377ffa7e8bf23c7b