Use nodepool-builder.logging.conf from puppet-nodepool
Generating a openstack-infra specific version of this file was
deprecated with change I6fdc3b622454b069f3ad2cf42da584d8df23a110 when
build logs was moved out of Python logging.

This means (post the depends-on) the file is exactly as kept in
puppet-nodepool.  Remove the system-config template and just use that.

Depends-On: https://review.openstack.org/586231
Change-Id: I2db37868be12c146678f1c8934db80050c788540