Install and configure ansible on bridge
There is a shared caching infrastructure in ansible now for inventory
and fact plugins. It needs to be configured so that our inventory access
isn't slow as dirt.

Unfortunately the copy of openstack.py in 2.6 is busted WRT to caching
because the internal API changed ... and we didn't have any test jobs
set up for it. This also includes a fixed copy of the plugin and
installs it into the a plugin dir.

Add a stub inventory file to suck in secrets

host_vars and group_vars are loaded from locations adjacent to inventory.
This adds a simple inventory entry for bridge.openstack.org that sets
ansible_connection to local. The more important thing is does is put a
static inventory in place so that the host_vars and group_vars from
secrets.

Change-Id: Ie92e5d7eac4b7e4060a4e07cb29c5a6f2a16ae18