Add a stub inventory file to suck in secrets
host_vars and group_vars are loaded from locations adjacent to inventory.
This adds a simple inventory entry for bridge.openstack.org that sets
ansible_connection to local. The more important thing is does is put a
static inventory in place so that the host_vars and group_vars from
secrets.

Change-Id: Ied2ac806bcc6858a6f6bb7861db0b48ba26b8908