Replace with_first_found with lookup first_found
The with_ directives are discouraged now in place of use of loop: and/or
lookup filters. In the case of with_first_found, it confuses people
because with_ directives are usually a loop, but in this case it's
the task is always executed once. Using the first_found filter makes it
clearer that this is occuring.

While we're in there, remove uses of 'static: no'. Since 2.0 includes
are dynamic by default, so these are not necessary.

Change-Id: Ie429d7614b2f3322a646f46a8117d4b6ae29f737