Use a when clause to make skip work with first_found
There are some issues with lookup('first_found') and skip: True. An
issue has been reported here:

  https://github.com/ansible/ansible/issues/43833

A workaround (at least I hope it is) is to do the lookup while setting a
variable, then add a when: to test that the variable isn't [] - which is
what first_found returns when skip: true.

Change-Id: If34d93886c702ea60d5669532c287ef7505ec214