Add exim role to base playbook
We want email to work.

Add a default value so that integration tests work - and update the
template so that if the value in the alias mapping is empty we don't
write out a half-formed alias.

Enable the epel repo on CentOS nodes in base-repos. This is done in
install_puppet.sh, but install_puppet.sh doesn't get run on ansible-only
nodes.

Change-Id: I68ad9f66c3b8672d9642c7764e50adac9cafdaf9