Turn on the future parser for eavesdrop.o.o
eavesdrop.o.o runs a number of different sites and bots, so on the one
hand it is a slightly complex service, but on the other hand these
modules are quite simple and similar to each other. Meetbot, accessbot,
and statusbot have tests and seem to work fine on puppet 4. Ptgbot does
not have tests, but that's why we should try this out sooner rather than
later so we have time to correct any issues before the PTG.

Change-Id: Id1af4b3ebf04f34c381eab5f818815734a296f7b