haproxy: don't manage rsyslog service
Ansible handles rsyslog, so rather than having both systems attempting
to manage the service, just let Ansible do it.  However, we still install
the haproxy rsyslog config file in Puppet for now.  The downside of that
is that an initial server build, or change to the config file, will not
cause rsyslog to reload.  That will need to be done manually until the
rest of this is ported to Ansible.

Change-Id: I19060fddf55e7e9f7030e49592f20a4a7faca429