Remove restricted and multiverse from ubuntu servers
restricted is supported software that is non-free.
multiverse is unsupported software that is non-free.

Use of software from either would be unacceptable on any Infra server,
so remove them from the sources.list files.

While we're in there, clean things up a little bit and add an arm file
for bionic.

Change-Id: I55a3b3d411e8a3496a4e6910baaf72f3c192e9d4