Generate roles documentation
Since we're building out roles in system-config now, generate
documentation.

Correct the reference names in the exim documentation to avoid
warnings and failure.

There is also a single unicode character in the exim readme.  For fun,
see if you can find it.

Because we currently don't override the default python2 build for the
tox docs environment, zuul-sphinx needs some help to open the file in
unicode mode and not pass byte-strings to the RST core, which
subsequently explodes in a long backtrace ... see Depends-On.  Note
will we will want a zuul-sphinx release because it's not co-installed
(however, see 593484 for manual testing).

A follow-on (I2c49a0b8bcda72385a9594f40a6b9db10403d688) will move doc
building to python3, but since the problem is identified we may as
well fix it ... zuul-jobs, and thus by exension zuul-sphinx, has a
python2 support policy.

Tested-By: https://review.openstack.org/593484
Depends-On: https://review.openstack.org/579474
Change-Id: I243a96bbd6d09560f8aa80b6345b90039422547a