Generate roles documentation
Since we're building out roles in system-config now, generate
documentation.

Correct the reference names in the exim documentation.

There is a unicode character in the exim documenation.  For fun, see
if you can find it.

Because we currently build documentation with python2, zuul-sphinx
needs some help to open the file in unicode mode and not pass
byte-strings to the RST core which subsequently explodes in a long
backtrace ... see Depends-On.  Note will we will want a zuul-sphinx
release because it's not co-installed.

A follow-on will move things to python3, but since the problem is
identified we may as well fix it.

Depends-On: https://review.openstack.org/#/c/579474/
Change-Id: I243a96bbd6d09560f8aa80b6345b90039422547a