roles/exim: remove "true" filter plugin
With the roles moved to top-level roles/ dir, when actually trying to
use system-config roles from a zuul job you end up with:

 ERROR Ansible plugin dir .../system-config/roles/exim/filter_plugins
   found adjacent to playbook ../system-config/roles/exim in non-trusted repo.

because system-config isn't a trusted repo.

These roles doing dual-duty as generic ansible roles (for infra
control-plane) and as zuul-jobs is probably the first of its type.  I
don't think we want system-config as a trusted repo if we can help it.

However, in this particular case, I don't believe this filter is
actually used in the role; thus remove it.  This avoids the problem
(for now).

Change-Id: Ice97f0c3c9f52b6bf9f48c7b16d577e555924034