Allow bridge playbook to run under zuul
Normally the bridge playbook runs as root on bridge.  In order to
allow zuul to bootstrap a bridge-like node in its tests while running
as the zuul user, add become: true to the playbook.  This will have
no effect on bridge itself, but will cause the playbook to behave
in the same manner in tests.

Also add the "users" role to bridge.  This is in the base playbook
and is therefore eventually run on bridge.  However it needs to also
be in the bridge playbook in order to bootstrap bridge correctly, as
the install-ansible role references groups which are created in the
users role.

Change-Id: If311914e9e632d8be855fff0a62528dd191bf1d0