Create an ansible group for bridge.o.o
This is a follow-on to I6a7fdd460fb472f0d3468eb080aebbb010931e11 where
it was suggested that the "admin" group owning Ansible's runtime
directories on bridge.o.o was useful to launch nodes without full sudo
permissions.

Restore this in a more fine-grained manner.  Create an ansible group
during Ansible installation in "install-ansible", and give access to
the runtime directories.  Modify this role to set a flag that is
picked up by the "users" role to add the group to users on hosts where
Ansible is installed.

Update testinfra to check for the group and permissions on bridge.o.o;
and conversely its absence on other hosts.

Change-Id: Id182a653b22af4c3fcb5333548d5974594005622