Install python2.7 on xenial nodes
Ubuntu xenial does not come with python2 by default. In order to
accomodate a transition from trusty nodes to xenial nodes that are
managed by ansible we want to use python2 on trusty and xenial. Then
when a group of nodes are fully xenialed we can force ansible to use
python3 instead.

Eventually we will have no trusty nodes and can default to using
python3 instead and just have to have a small number of exceptions for
centos.

Change-Id: If1d97e25069d6ed5012c147024aad4d921febfc8