Add system-config-run-eavesdrop
Add a job which runs testinfra for the eavesdrop server.  When we
have a per-hostgroup playbook, we will add it to this job too.

The puppet group is removed from the run-base job because the
groups.yaml file is now used to construct groups (as it does
in production) and will construct the group correctly.

The testinfra iptables module should use '-w' to support parallel
execution.  A temporary copy which does this is added here until
upstream can be updated.

Change-Id: I1a7bab5c14b0da22393ab568000d0921c28675aa