WIP: Manage clouds.yaml files in ansible
This manages the clouds.yaml files in ansible so that we can get them
updated automatically on bridge.openstack.org (which does not puppet).

Note the new .j2 files need to be converted to jinja with the correct
var lookups and we need to clean up the old side of things.

Change-Id: I2071f2593f57024bc985e18eaf1ffbf6f3d38140