Run puppet apply test serially on centos
There appears to be a race running the ansible synchronize (rsync under
the hood) top copy puppet modules for multiple puppet applies at the
same time on CentOS7. Running this in parallel appears safe on Ubuntu
and does save quite a bit of job runtime.

Workaround this by running the apply test serially on CentOS only.

Change-Id: Icd0836db215c0b417989d38994a378a705bbc62b