Revert "Fix ansible forks env variable"
And revert "Set Ansible forks to 50"

This doesn't seem to have helped, and may have made the run longer.
I suspect a problem with the env var, but let's revert back to the
old value and mechanism (cli flag) to re-establish a baseline,
then we'll change the value of the cli flag.

This reverts commit 84199095716da416849ed4a2649ec8a2c878609d.
This reverts commit 97d8f9d0bfaec24413f134fe252d4011fe9e36d4.

Change-Id: I825b2b3db26ce6dd7d70fcc8b33e70b511eb52db