Don't set security groups in openstackci-packethost
This account is an admin account and sees every project's default
security group. This leads to:

  FAILED! => {"changed": false, "msg": "Multiple matches found for default"}

When attempting to set the properties of the default security group for
this account. There doesn't appear to be a good way to filter the other
default security groups out currently so avoid setting them for now.

Change-Id: I9a8cc7d59c0295caa71bf107b9b78745a4617981