Serialize puppet on afs servers properly
We were matching afs* as a glob to serialize puppet runs on afs servers.
This was fine until we added afs-client and afsadmin groups to our
inventory which matched afs*. These groups included many nodes including
our mirror nodes and zuul executors all of which were running puppet
serially which is slow.

Fix this by explicitly using the afs and afsdb groups instead of a glob.

Change-Id: If21bbc48b19806343617e18fb03416c642e00ed2