Run base playbook with strategy free
The base playbook plays that run on every node do so independent of
other nodes. We set things up like exim and unbound and user accounts
all of which can run as quickly as possible on each host rather than in
lockstep across all of the hosts.

This change should make better use time slicing across ansible forks as
we'll go as quickly as possible for each thread rather than waiting
periodically.

Change-Id: Ic2ff5722ac739908351b7c2f60bf16e14dca4a30
Depends-On: https://review.openstack.org/600861