Remove !ci-backup play
Let's abandon the idea that we'll treat the backup server specially.
As long as we allow *any* automated remote access via ansible, we
have opened the door to potential compromise of the backup systems
if bridge is compromised.  Rather than pretending that this separation
gives us any benefit, remove it.

Change-Id: I751060dc05918c440374e80ffb483d948f048f36