Allow project website volume path to be overridden
We've created a /afs/.openstack.org/project/starlingx.io volume for
starlingx to host specs and docs and whatever else. Currently they are
trying to set up docs.starlingx.io as a vhost to serve docs. This Means
the vhost name and volume path differ. Allow for specifying a volume
name that is different than the volume path in this case. The idea here
is to avoid needing to create separate afs volumes for different (but
related) content.

Change-Id: Iea6d38096f2b4c90feab56289003af29fd5c224d